+++ rust/candid_parser/tests/assets/ok/cyclic.js
@@ -1,3 +1,19 @@
+import { IDL } from '@dfinity/candid';
+
+export const A = IDL.Rec();
+export const C = A;
+export const B = IDL.Opt(C);
+A.fill(IDL.Opt(B));
+export const Z = A;
+export const Y = Z;
+export const X = Y;
+
+export const idlService = IDL.Service({
+  'f' : IDL.Func([A, B, C, X, Y, Z], [], []),
+});
+
+export const idlInitArgs = [];
+
 export const idlFactory = ({ IDL }) => {
   const A = IDL.Rec();
   const C = A;
@@ -6,6 +22,8 @@ export const idlFactory = ({ IDL }) => {
   const Z = A;
   const Y = Z;
   const X = Y;
+  
   return IDL.Service({ 'f' : IDL.Func([A, B, C, X, Y, Z], [], []) });
 };
+
 export const init = ({ IDL }) => { return []; };